+++ example/main.dart
@@ -165,7 +165,7 @@ class BookMapper implements GlobalResourceMapper<Book> {
     final String title = reader.require(SchemaBook.name);
     final String authorId = reader.require(
       SchemaBook.author,
-      iriTermDeserializer: _authorIdMapper,
+      deserializer: _authorIdMapper,
     );
     final DateTime published = reader.require(SchemaBook.datePublished);
     final ISBN isbn = reader.require(SchemaBook.isbn);
@@ -188,7 +188,7 @@ class BookMapper implements GlobalResourceMapper<Book> {
   }
 
   @override
-  (IriTerm, List<Triple>) toRdfResource(
+  (IriTerm, Iterable<Triple>) toRdfResource(
     Book resource,
     SerializationContext context, {
     RdfSubject? parentSubject,
@@ -201,7 +201,7 @@ class BookMapper implements GlobalResourceMapper<Book> {
         .addValue(
           SchemaBook.author,
           resource.authorId,
-          iriTermSerializer: _authorIdMapper,
+          serializer: _authorIdMapper,
         )
         .addValue(SchemaBook.datePublished, resource.published)
         .addValue(SchemaBook.isbn, resource.isbn)
@@ -243,7 +243,7 @@ class ChapterMapper implements LocalResourceMapper<Chapter> {
   }
 
   @override
-  (BlankNodeTerm, List<Triple>) toRdfResource(
+  (BlankNodeTerm, Iterable<Triple>) toRdfResource(
     Chapter resource,
     SerializationContext context, {
     RdfSubject? parentSubject,
@@ -299,6 +299,9 @@ class ISBNMapper implements IriTermMapper<ISBN> {
 /// and RDF terms for iri terms of type Rating.
 class RatingMapper implements LiteralTermMapper<Rating> {
   const RatingMapper();
+
+  @override
+  IriTerm? get datatype => null;
 
   @override
   Rating fromRdfTerm(